requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.13.4
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dotenv==0.9.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.2
frozenlist==1.6.0
fsspec==2025.5.1
future==1.0.0
google-auth==2.40.1
google-genai==1.14.0
greenlet==3.2.2
griffe==1.7.3
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language-data==1.3.0
llama-cloud==0.1.21
llama-cloud-services==0.6.15
llama-index==0.12.35
llama-index-agent-openai==0.4.7
llama-index-cli==0.4.1
llama-index-core==0.12.35
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-google-genai==0.1.12
llama-index-llms-ollama==0.5.4
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.12
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
murmurhash==1.0.13
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.3.0
ollama==0.4.8
openai==1.78.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pke @ git+https://github.com/boudinfl/pke.git@69871ffdb720b83df23684fea53ec8776fd87e63
platformdirs==4.3.8
pluggy==1.6.0
portalocker==2.10.1
preshed==3.0.10
propcache==0.3.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.1
pyparsing==3.2.3
pypdf==5.4.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.40
srsly==2.5.1
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typer==0.16.0
typing-extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
unidecode==1.4.0
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
